Fix 500s under concurrency: DescribeExec is not pooler-safe either
Loading /explore fired three requests at once and two came back 500. The same
endpoints returned 200 every time when called one after another, which is why
every check so far had passed.

QueryExecModeDescribeExec sends Parse+Describe, then Bind+Execute — two round
trips. A pooler in transaction mode treats those as two transactions and is
free to hand the connection to another client in between, at which point the
unnamed statement is gone. With nothing else competing for connections it never
happens; with a real page load it happens immediately.

So all three of the obvious modes fail on a transaction-mode pooler for
different reasons, and only one is actually safe:

  CacheStatement  named prepared statements -> 42P05 on connection reuse
  DescribeExec    two round trips -> breaks only under concurrency
  Exec            one round trip, holds nothing -> correct

Exec was tried earlier and reverted because it broke four update tests. That
was the right signal read the wrong way: under Exec pgx infers parameter types
itself instead of asking the server, so a parameter whose type is not obvious
from context needs an explicit cast. There was exactly one such parameter.

  update.List passed []uuid.UUID to `= ANY($1)`. pgx has no encode plan for a
  uuid slice — "unable to encode []uuid.UUID ... for unknown type (OID 0)" —
  and a ::uuid[] cast alone does not help, because the failure is client-side
  encoding before any SQL is sent. It now passes []string with ::uuid[].

Three more parameters stand alone as predicates with no column to infer from,
and are now cast explicitly rather than left to luck: the includeDrafts and
includeInternal booleans, the super-admin boolean in the society listing, and
the fallback rate in COALESCE.

Verified against the real Supabase pooler with the production image: 10 rounds
of the browser's three concurrent requests, then 30 parallel hits on the
endpoint that failed most often. 30/30 returned 200 and the container logged no
errors. The full suite is green.

+++ internal/maintenance/maintenance.go
@@ -71,7 +71,7 @@ func (s *Store) List(ctx context.Context, societyID uuid.UUID) ([]Rate, float64,
 
 	rows, err := s.db.Query(ctx, `
 		SELECT p.phase,
-		       COALESCE(r.rate_per_sqft, $2) AS rate,
+		       COALESCE(r.rate_per_sqft, $2::numeric) AS rate,
 		       COALESCE(r.id, '00000000-0000-0000-0000-000000000000'::uuid),
 		       COALESCE(r.updated_at, now()),
 		       count(*)::int AS plots,
